Fix type of WKSResearchModule.RatePercentages

+++ FFXIVClientStructs/FFXIV/Client/Game/WKS/WKSResearchModule.cs
@@ -10,7 +10,7 @@ public unsafe partial struct WKSResearchModule {
     [FieldOffset(0x08), FixedSizeArray] internal FixedSizeArray55<ushort> _analysis; // 11x toolClass * 5x type
     [FieldOffset(0x76), FixedSizeArray] internal FixedSizeArray11<byte> _currentStages;
     [FieldOffset(0x81), FixedSizeArray] internal FixedSizeArray11<byte> _unlockedStages;
-    [FieldOffset(0x8C), FixedSizeArray] internal FixedSizeArray2<byte> _ratePercentages;
+    [FieldOffset(0x8C), FixedSizeArray] internal FixedSizeArray2<ushort> _ratePercentages;
     [FieldOffset(0x90)] public bool IsLoaded;
 
     [MemberFunction("E8 ?? ?? ?? ?? BA ?? ?? ?? ?? 0F B7 D8 48 8B CF E8 ?? ?? ?? ?? 48 8B 4C 24")]
